pyproject: avoid broken docutils release
Poetry fails with:

Package docutils (0.21.post1) not found.

because of a broken docutils release. Avoid
the 0.21 release, and run poetry update for
a more recent lock file.

The pre-commit fails with:

RuntimeError: The Poetry configuration is invalid:
            - data.extras.pipfile_deprecated_finder[2] must match pattern ^[a-zA-Z-_.0-9]+$

so update to isort 5.11.5 that contains a poetry
hotfix, and seems to work.

Some other issue with black, so update that
to 22.12.0 which also seems to work.